update stage 4, new package payment

+++ cinema/cinema-core/src/main/java/com/nongviet201/cinema/core/model/entity/bill/Bill.java
@@ -19,7 +19,7 @@ public class Bill {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     int id;
     boolean status;
-    BigDecimal totalPrice;
+    long totalPrice;
     LocalDate createAt;
     LocalDate updateAt;
