+++ cinema/cinema-core/src/main/java/com/nongviet201/cinema/core/model/entity/bill/BillCombo.java
@@ -17,7 +17,7 @@ public class BillCombo {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     int id;
     int quantity;
-    BigDecimal price;
+    long price;
 
     @ManyToOne
     @JoinColumn(name= "bill_id")
