+++ cinema/cinema-core/src/main/java/com/nongviet201/cinema/core/model/entity/bill/Combo.java
@@ -21,5 +21,5 @@ public class Combo {
     String description;
     String image;
     boolean status;
-    BigDecimal price;
+    long price;
 }
