+++ cinema/cinema-core/src/main/java/com/nongviet201/cinema/core/model/entity/bill/Coupon.java
@@ -17,7 +17,7 @@ public class Coupon {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     int id;
-    BigDecimal discountPrice;
+    long discountPrice;
     int limitAmount;
     String code;
     LocalDate startDate;
